Merge remote branch 'origin/master' into js

Conflicts:
	demo/scenes/colladasmall.csv
	liboh/src/HostedObject.cpp
	libproxyobject/plugins/ogre/OgreSystemMouseHandler.cpp

+++ liboh/src/SessionManager.cpp
@@ -664,7 +664,7 @@ void SessionManager::handleSessionMessage(Sirikata::Protocol::Object::ObjectMess
             mObjectConnections.handleConnectError(obj);
         }
         else {
-            OH_LOG(error,"Unknown connection response code: " << conn_resp.response());
+            OH_LOG(error,"Unknown connection response code: " << (int)conn_resp.response());
         }
     }
